roe-1287 Commented work in progress code
This code changes test functionality and is preventing a push to remote
branch. Test will need to be updated when code is uncommented.
I have not fixed the test in this commit as the purpose of the push is
to get feedback on the working branch. This change may not be the best
approach. If change is kept then test will be fixed accordindly in a
future commit