roe-1731: Fix session mapping for historic trustee
Historic trustee wasn't being saved to the session properly. Due to the
type being used to determine whether former trustee was individual or LE
the LE was never being selected and thus the corporate name would not be
save to the session. THe name wasn't in the session so I couldn't
display it on the trust involved page

- use TrusteeType throughout historic trustee could to fix mapping
  issue