roe-1566: create and modify util mehtods
- create getTrustArray which returns the array of trust data from the
  current application object in the session

- modify saveTrustInApp
 A bug was discovered on the check your answers page during this story
due to now managing multiple trusts. If you were to add 3 trusts to an
application they would initally be displayed as trust1, trust2, trust3
on check your answers. But if you navigate back and modify trust2 then
they now appear on CYA as trust1, trust3, trust2.
The data is still in tact but feels quite jarring as a use.

The code has now been modified to update the trust data at it's current
index rather than pulling it out, updating and appending to the end of
the arrayr